refactor(deps): move Pipfile/lockfile bridge functions to dependencies.py
T_E.2: Per the T_E.1 design + sign-off, move 6 Pipfile/lockfile
bridge functions plus BAD_PACKAGES from pipenv/utils/requirements.py
to pipenv/utils/dependencies.py (the canonical home). Rename the
module-level add_index_to_pipfile to add_index_to_pipfile_with_trust_check
to disambiguate from Project.add_index_to_pipfile per T_E.1 §3 sign-off.

Moved:
  import_requirements
  add_index_to_pipfile -> add_index_to_pipfile_with_trust_check
  requirement_from_lockfile
  requirements_from_lockfile
  requirement_from_pipfile
  requirements_from_pipfile
  BAD_PACKAGES

Caller imports migrated in the same PR per T_C.3 §9 / T_E.1
sign-off — no backwards-compat shim. After this commit
requirements.py contains only the redact_* fork pair (which stays
per T_E.1 §6 sign-off); T_E.7 will optionally rename the file to
redact.py.
