+++ pipenv/routines/clean.py
@@ -4,9 +4,9 @@ from pipenv.patched.pip._internal.build_env import get_runnable_pip
 from pipenv.routines.context import RoutineContext
 from pipenv.routines.lock import do_lock
 from pipenv.utils import console, err
+from pipenv.utils.dependencies import BAD_PACKAGES
 from pipenv.utils.processes import run_command
 from pipenv.utils.project import ensure_project
-from pipenv.utils.requirements import BAD_PACKAGES
 from pipenv.utils.shell import project_python
