+++ pipenv/routines/graph.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from pipenv import exceptions
 from pipenv.patched.pip._vendor.rich.markup import escape as rich_escape
 from pipenv.utils import console, err
+from pipenv.utils.dependencies import BAD_PACKAGES
 from pipenv.utils.processes import run_command
-from pipenv.utils.requirements import BAD_PACKAGES
 
 
 def do_graph(project, bare=False, json=False, json_tree=False, reverse=False):
